Consolidate test files to one per module with honest names
Eliminate the misleading _more/_coverage/_extended/_new suffixes so each
source module maps to a single canonical test file:

Renamed (no base existed):
- client/runtime/test_async_runner_more.py -> test_async_runner.py
- spec_guard/test_mappings_coverage.py    -> test_mappings.py

Merged sibling into base, then deleted sibling (all tests preserved):
- auth/test_oauth_coverage.py            -> test_oauth.py
- client/test_protocol_client_extended.py -> test_protocol_client.py
- fuzz_engine/runtime/test_watchdog_extended.py -> test_watchdog.py
- fuzz_engine/strategy/test_aggressive_tool_strategy_more.py
                                          -> test_aggressive_tool_strategy.py
- spec_guard/test_spec_checks_extended.py + _new.py -> test_spec_checks.py
- test_audit_fix_coverage.py              -> test_audit_fixes.py

Pure reorganization: same 2467 tests, identical 95.385% coverage, zero
per-file regression, ruff clean, green in both orderings.
