+++ tests/unit/spec_guard/test_spec_checks.py
@@ -4,6 +4,31 @@
 from __future__ import annotations
 
 from mcp_fuzzer.spec_guard import spec_checks
+from mcp_fuzzer.spec_guard.check_ids import CheckID
+from mcp_fuzzer.spec_guard.spec_checks import (
+    check_tool_schema_fields,
+    check_tool_result_content,
+    check_logging_notification,
+    check_resources_list,
+    check_resources_read,
+    check_resource_templates_list,
+    check_prompts_list,
+    check_prompts_get,
+    check_sse_event_text,
+    check_tools_list,
+    check_tasks_list,
+    check_roots_list,
+    check_create_message_result,
+    check_elicit_result,
+    check_task_result,
+    check_cancelled_notification,
+    check_completion_complete,
+    check_list_changed_notification,
+    check_progress_notification,
+    check_resources_updated_notification,
+    check_subscribe_result,
+    check_unsubscribe_result,
+)
 
 
 def test_check_tool_schema_fields_reports_errors():
@@ -179,3 +204,771 @@ def test_check_prompts_get_reports_missing_or_bad_messages():
 def test_check_sse_event_text_no_warnings_when_data_present():
     checks = spec_checks.check_sse_event_text("data: ok\nretry: 200\nid: evt1\n")
     assert checks == []
+
+
+# ===========================================================================
+# Merged from test_spec_checks_extended.py
+# ===========================================================================
+
+
+class TestCheckToolSchemaFields:
+    """Test check_tool_schema_fields function."""
+
+    def test_no_input_schema(self):
+        """Test tool with no inputSchema."""
+        result = check_tool_schema_fields({})
+        assert result == []
+
+    def test_non_dict_input_schema(self):
+        """Test tool with non-dict inputSchema."""
+        result = check_tool_schema_fields({"inputSchema": "not a dict"})
+        assert result == []
+
+    def test_invalid_schema_type(self):
+        """Test tool with non-string $schema."""
+        result = check_tool_schema_fields({"inputSchema": {"$schema": 123}})
+        assert len(result) == 1
+        assert result[0]["id"] == "tool-schema-$schema"
+
+    def test_invalid_defs_type(self):
+        """Test tool with non-object $defs."""
+        result = check_tool_schema_fields({"inputSchema": {"$defs": "not a dict"}})
+        assert len(result) == 1
+        assert result[0]["id"] == "tool-schema-$defs"
+
+    def test_invalid_additional_properties(self):
+        """Test tool with invalid additionalProperties."""
+        result = check_tool_schema_fields(
+            {"inputSchema": {"additionalProperties": "invalid"}}
+        )
+        assert len(result) == 1
+        assert result[0]["id"] == "tool-schema-additional-properties"
+
+    def test_valid_additional_properties_bool(self):
+        """Test tool with valid boolean additionalProperties."""
+        result = check_tool_schema_fields(
+            {"inputSchema": {"additionalProperties": True}}
+        )
+        assert result == []
+
+    def test_valid_additional_properties_dict(self):
+        """Test tool with valid dict additionalProperties."""
+        result = check_tool_schema_fields(
+            {"inputSchema": {"additionalProperties": {"type": "string"}}}
+        )
+        assert result == []
+
+
+class TestCheckToolResultContent:
+    """Test check_tool_result_content function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_tool_result_content("not a dict")
+        assert result == []
+
+    def test_no_content_key(self):
+        """Test result without content key."""
+        result = check_tool_result_content({"other": "data"})
+        assert result == []
+
+    def test_non_array_content(self):
+        """Test result with non-array content."""
+        result = check_tool_result_content({"content": "not an array"})
+        assert len(result) == 1
+        assert result[0]["id"] == "tools-content-array"
+
+    def test_empty_content_array(self):
+        """Test result with empty content array."""
+        result = check_tool_result_content({"content": []})
+        assert len(result) == 1
+        assert result[0]["id"] == "tools-content-empty"
+
+    def test_non_dict_content_item(self):
+        """Test result with non-dict content item."""
+        result = check_tool_result_content({"content": ["not a dict"]})
+        assert any(c["id"] == "tools-content-item" for c in result)
+
+    def test_content_item_missing_type(self):
+        """Test content item missing type."""
+        result = check_tool_result_content({"content": [{"text": "hello"}]})
+        assert any(c["id"] == "tools-content-type" for c in result)
+
+    def test_text_content_missing_text(self):
+        """Test text content missing text field."""
+        result = check_tool_result_content({"content": [{"type": "text"}]})
+        assert any(c["id"] == "tools-content-text" for c in result)
+
+    def test_text_content_empty_text(self):
+        """Test text content with empty text field."""
+        result = check_tool_result_content({"content": [{"type": "text", "text": ""}]})
+        assert any(c["id"] == "tools-content-text" for c in result)
+
+    def test_image_content_missing_data(self):
+        """Test image content missing data field."""
+        result = check_tool_result_content(
+            {"content": [{"type": "image", "mimeType": "image/png"}]}
+        )
+        assert any(c["id"] == "tools-content-image-data" for c in result)
+
+    def test_image_content_missing_mimetype(self):
+        """Test image content missing mimeType field."""
+        result = check_tool_result_content(
+            {"content": [{"type": "image", "data": "base64..."}]}
+        )
+        assert any(c["id"] == "tools-content-image-mime" for c in result)
+    def test_audio_content_missing_data(self):
+        """Test audio content missing data field."""
+        result = check_tool_result_content(
+            {"content": [{"type": "audio", "mimeType": "audio/mp3"}]}
+        )
+        assert any(c["id"] == "tools-content-audio-data" for c in result)
+
+    def test_audio_content_missing_mimetype(self):
+        """Test audio content missing mimeType field."""
+        result = check_tool_result_content(
+            {"content": [{"type": "audio", "data": "base64..."}]}
+        )
+        assert any(c["id"] == "tools-content-audio-mime" for c in result)
+
+    def test_resource_content_non_dict_resource(self):
+        """Test resource content with non-dict resource."""
+        result = check_tool_result_content(
+            {"content": [{"type": "resource", "resource": "not a dict"}]}
+        )
+        assert any(c["id"] == "tools-content-resource" for c in result)
+
+    def test_resource_content_missing_uri(self):
+        """Test resource content missing uri."""
+        result = check_tool_result_content(
+            {
+                "content": [
+                    {
+                        "type": "resource",
+                        "resource": {"mimeType": "text/plain", "text": "data"},
+                    }
+                ]
+            }
+        )
+        assert any(c["id"] == "tools-content-resource-uri" for c in result)
+
+    def test_resource_content_missing_mimetype(self):
+        """Test resource content missing mimeType is allowed."""
+        result = check_tool_result_content(
+            {
+                "content": [
+                    {
+                        "type": "resource",
+                        "resource": {"uri": "file://test", "text": "data"},
+                    }
+                ]
+            }
+        )
+        assert not any(c["id"] == "tools-content-resource-mime" for c in result)
+
+    def test_resource_content_missing_body(self):
+        """Test resource content missing text or blob."""
+        result = check_tool_result_content(
+            {
+                "content": [
+                    {
+                        "type": "resource",
+                        "resource": {
+                            "uri": "file://test",
+                            "mimeType": "text/plain",
+                        },
+                    }
+                ]
+            }
+        )
+        assert any(c["id"] == "tools-content-resource-body" for c in result)
+
+    def test_unknown_content_type(self):
+        """Test unknown content type generates warning."""
+        result = check_tool_result_content({"content": [{"type": "custom_type"}]})
+        assert any(c["id"] == "tools-content-unknown-type" for c in result)
+        # Should be a warning, not a failure
+        assert any(c.get("status") == "WARN" for c in result)
+
+    def test_is_error_without_text_message(self):
+        """Test isError=true without text error message."""
+        result = check_tool_result_content(
+            {
+                "isError": True,
+                "content": [
+                    {"type": "image", "data": "base64", "mimeType": "image/png"}
+                ],
+            }
+        )
+        assert any(c["id"] == "tools-error-text" for c in result)
+
+    def test_is_error_with_text_message(self):
+        """Test isError=true with proper text error message."""
+        result = check_tool_result_content(
+            {"isError": True, "content": [{"type": "text", "text": "Error occurred"}]}
+        )
+        # Should not have tools-error-text failure
+        assert not any(c["id"] == "tools-error-text" for c in result)
+
+
+class TestCheckLoggingNotification:
+    """Test check_logging_notification function."""
+
+    def test_no_params(self):
+        """Test payload without params."""
+        result = check_logging_notification({})
+        assert any(c["id"] == "logging-params-missing" for c in result)
+
+    def test_params_none(self):
+        """Test payload with None params."""
+        result = check_logging_notification({"params": None})
+        assert any(c["id"] == "logging-params-missing" for c in result)
+
+    def test_non_dict_params(self):
+        """Test payload with non-dict params."""
+        result = check_logging_notification({"params": "not a dict"})
+        assert len(result) == 1
+        assert result[0]["id"] == "logging-params-type"
+
+    def test_invalid_level_type(self):
+        """Test payload with non-string level."""
+        result = check_logging_notification({"params": {"level": 123}})
+        assert any(c["id"] == "logging-level-type" for c in result)
+
+    def test_invalid_message_type(self):
+        """Test payload with non-string logger."""
+        result = check_logging_notification({"params": {"logger": 123}})
+        assert any(c["id"] == "logging-logger-type" for c in result)
+        assert any(c["id"] == "logging-level-missing" for c in result)
+        assert any(c["id"] == "logging-data-missing" for c in result)
+
+    def test_valid_logging_notification(self):
+        """Test valid logging notification."""
+        result = check_logging_notification(
+            {"params": {"level": "INFO", "data": "Test"}}
+        )
+        assert result == []
+
+
+class TestCheckResourcesList:
+    """Test check_resources_list function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_resources_list("not a dict")
+        assert result == []
+
+    def test_missing_resources(self):
+        """Test result missing resources key."""
+        result = check_resources_list({})
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-list-missing"
+
+    def test_non_array_resources(self):
+        """Test result with non-array resources."""
+        result = check_resources_list({"resources": "not an array"})
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-list-type"
+
+    def test_non_dict_resource_item(self):
+        """Test resource item that is not a dict."""
+        result = check_resources_list({"resources": ["not a dict"]})
+        assert any(c["id"] == "resources-list-item" for c in result)
+
+    def test_resource_missing_uri(self):
+        """Test resource missing uri."""
+        result = check_resources_list({"resources": [{"name": "test"}]})
+        assert any(c["id"] == "resources-list-uri" for c in result)
+
+    def test_resource_missing_name(self):
+        """Test resource missing name."""
+        result = check_resources_list({"resources": [{"uri": "file://test"}]})
+        assert any(c["id"] == "resources-list-name" for c in result)
+
+    def test_valid_resources(self):
+        """Test valid resources list."""
+        result = check_resources_list(
+            {"resources": [{"uri": "file://test", "name": "Test"}]}
+        )
+        assert result == []
+
+
+class TestCheckResourcesRead:
+    """Test check_resources_read function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_resources_read("not a dict")
+        assert result == []
+
+    def test_missing_contents(self):
+        """Test result missing contents key."""
+        result = check_resources_read({})
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-read-missing"
+
+    def test_non_array_contents(self):
+        """Test result with non-array contents."""
+        result = check_resources_read({"contents": "not an array"})
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-read-type"
+
+    def test_empty_contents(self):
+        """Test result with empty contents array."""
+        result = check_resources_read({"contents": []})
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-read-empty"
+
+    def test_non_dict_content_item(self):
+        """Test content item that is not a dict."""
+        result = check_resources_read({"contents": ["not a dict"]})
+        assert any(c["id"] == "resources-read-item" for c in result)
+
+    def test_content_missing_uri(self):
+        """Test content missing uri."""
+        result = check_resources_read({"contents": [{"text": "data"}]})
+        assert any(c["id"] == "resources-read-uri" for c in result)
+
+    def test_content_missing_body(self):
+        """Test content missing text and blob."""
+        result = check_resources_read({"contents": [{"uri": "file://test"}]})
+        assert any(c["id"] == "resources-read-body" for c in result)
+
+    def test_valid_contents(self):
+        """Test valid contents."""
+        result = check_resources_read(
+            {"contents": [{"uri": "file://test", "text": "data"}]}
+        )
+        assert result == []
+
+
+class TestCheckResourceTemplatesList:
+    """Test check_resource_templates_list function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_resource_templates_list("not a dict")
+        assert result == []
+
+    def test_no_resource_templates(self):
+        """Test result without resourceTemplates key."""
+        result = check_resource_templates_list({})
+        assert any(c["id"] == "resources-templates-missing" for c in result)
+
+    def test_non_array_templates(self):
+        """Test result with non-array resourceTemplates."""
+        result = check_resource_templates_list(
+            {"resourceTemplates": "not an array"}
+        )
+        assert len(result) == 1
+        assert result[0]["id"] == "resources-templates-type"
+
+    def test_non_dict_template_item(self):
+        """Test template item that is not a dict."""
+        result = check_resource_templates_list(
+            {"resourceTemplates": ["not a dict"]}
+        )
+        assert any(c["id"] == "resources-templates-item" for c in result)
+
+    def test_template_missing_uri_template(self):
+        """Test template missing uriTemplate."""
+        result = check_resource_templates_list(
+            {"resourceTemplates": [{"name": "test"}]}
+        )
+        assert any(c["id"] == "resources-templates-uri" for c in result)
+
+    def test_valid_templates(self):
+        """Test valid templates."""
+        result = check_resource_templates_list(
+            {"resourceTemplates": [{"name": "tmpl", "uriTemplate": "file://template"}]}
+        )
+        assert result == []
+
+
+class TestCheckPromptsList:
+    """Test check_prompts_list function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_prompts_list("not a dict")
+        assert result == []
+
+    def test_missing_prompts(self):
+        """Test result missing prompts key."""
+        result = check_prompts_list({})
+        assert len(result) == 1
+        assert result[0]["id"] == "prompts-list-missing"
+
+    def test_non_array_prompts(self):
+        """Test result with non-array prompts."""
+        result = check_prompts_list({"prompts": "not an array"})
+        assert len(result) == 1
+        assert result[0]["id"] == "prompts-list-type"
+
+
+class TestCheckPromptsGet:
+    """Test check_prompts_get function."""
+
+    def test_non_dict_result(self):
+        """Test with non-dict result."""
+        result = check_prompts_get("not a dict")
+        assert result == []
+
+    def test_missing_messages(self):
+        """Test result missing messages key."""
+        result = check_prompts_get({})
+        assert len(result) == 1
+        assert result[0]["id"] == "prompts-get-missing"
+
+    def test_non_array_messages(self):
+        """Test result with non-array messages."""
+        result = check_prompts_get({"messages": "not an array"})
+        assert len(result) == 1
+        assert result[0]["id"] == "prompts-get-type"
+
+
+class TestCheckSseEventText:
+    """Test check_sse_event_text function."""
+
+    def test_valid_event(self):
+        """Test valid SSE event."""
+        result = check_sse_event_text("data: {\"foo\": \"bar\"}")
+        assert result == []
+
+    def test_missing_data_prefix(self):
+        """Test SSE event without data prefix."""
+        result = check_sse_event_text("{\"foo\": \"bar\"}")
+        # Should warn about missing data field
+        assert any(c.get("status") == "WARN" for c in result)
+
+
+def test_check_task_result_rejects_bool_fields():
+    checks = check_task_result(
+        {
+            "taskId": "task-1",
+            "status": "running",
+            "createdAt": "2024-01-01T00:00:00Z",
+            "lastUpdatedAt": "2024-01-01T00:00:00Z",
+            "ttl": True,
+            "pollInterval": False,
+        }
+    )
+    ids = {check["id"] for check in checks}
+    assert "task-ttl" in ids
+    assert "task-poll-interval" in ids
+
+
+def test_spec_at_least_handles_invalid_version(monkeypatch):
+    monkeypatch.setenv("MCP_SPEC_SCHEMA_VERSION", "0000")
+    assert spec_checks._spec_at_least("9999-01-01") is False
+
+
+def test_tool_schema_icons_and_execution_validation():
+    ids = {check["id"] for check in check_tool_schema_fields({"icons": "nope"})}
+    assert "tool-icons-type" in ids
+
+    ids = {
+        check["id"]
+        for check in check_tool_schema_fields({"icons": ["bad", {"src": ""}]})
+    }
+    assert "tool-icon-item" in ids
+    assert "tool-icon-src" in ids
+
+    ids = {check["id"] for check in check_tool_schema_fields({"execution": "nope"})}
+    assert "tool-execution-type" in ids
+
+    ids = {
+        check["id"]
+        for check in check_tool_schema_fields({"execution": {"taskSupport": "maybe"}})
+    }
+    assert "tool-execution-task-support" in ids
+
+
+def test_tools_and_tasks_list_shape_errors():
+    ids = {check["id"] for check in check_tools_list({})}
+    assert "tools-list-missing" in ids
+    ids = {check["id"] for check in check_tools_list({"tools": "nope"})}
+    assert "tools-list-type" in ids
+    ids = {check["id"] for check in check_tools_list({"tools": ["bad"]})}
+    assert "tools-list-item" in ids
+
+    ids = {check["id"] for check in check_tasks_list({})}
+    assert "tasks-list-missing" in ids
+    ids = {check["id"] for check in check_tasks_list({"tasks": "nope"})}
+    assert "tasks-list-type" in ids
+
+
+def test_roots_sampling_and_elicitation_validation():
+    ids = {
+        check["id"]
+        for check in check_roots_list({"roots": ["bad", {"uri": ""}]})
+    }
+    assert "roots-list-item" in ids
+    assert "roots-list-uri" in ids
+
+    ids = {
+        check["id"]
+        for check in check_create_message_result(
+            {"model": "", "role": 123, "content": "bad", "stopReason": 5}
+        )
+    }
+    assert "sampling-model" in ids
+    assert "sampling-role" in ids
+    assert "sampling-content" in ids
+    assert "sampling-stop-reason" in ids
+
+    ids = {
+        check["id"]
+        for check in check_elicit_result({"action": "unknown", "content": "bad"})
+    }
+    assert "elicitation-action" in ids
+    assert "elicitation-content" in ids
+
+
+# ===========================================================================
+# Merged from test_spec_checks_new.py
+# ===========================================================================
+
+
+# ---------------------------------------------------------------------------
+# completion/complete
+# ---------------------------------------------------------------------------
+
+
+def test_completion_complete_valid():
+    result = {"completion": {"values": ["foo", "bar"], "hasMore": False, "total": 2}}
+    assert check_completion_complete(result) == []
+
+
+def test_completion_complete_minimal():
+    result = {"completion": {"values": []}}
+    assert check_completion_complete(result) == []
+
+
+def test_completion_complete_not_dict():
+    assert check_completion_complete("string") == []
+
+
+def test_completion_complete_missing_completion_key():
+    checks = check_completion_complete({})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_MISSING in ids
+
+
+def test_completion_complete_completion_not_dict():
+    checks = check_completion_complete({"completion": "bad"})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_TYPE in ids
+
+
+def test_completion_complete_missing_values():
+    checks = check_completion_complete({"completion": {}})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_VALUES_MISSING in ids
+
+
+def test_completion_complete_values_not_array():
+    checks = check_completion_complete({"completion": {"values": "bad"}})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_VALUES_TYPE in ids
+
+
+def test_completion_complete_values_non_string_item():
+    checks = check_completion_complete({"completion": {"values": [1, "ok"]}})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_VALUES_ITEM in ids
+
+
+def test_completion_complete_has_more_non_bool():
+    checks = check_completion_complete(
+        {"completion": {"values": [], "hasMore": "yes"}}
+    )
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_HAS_MORE_TYPE in ids
+
+
+def test_completion_complete_total_non_int():
+    checks = check_completion_complete(
+        {"completion": {"values": [], "total": 3.5}}
+    )
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_TOTAL_TYPE in ids
+
+
+def test_completion_complete_total_bool_rejected():
+    checks = check_completion_complete({"completion": {"values": [], "total": True}})
+    ids = [c["id"] for c in checks]
+    assert CheckID.COMPLETION_TOTAL_TYPE in ids
+
+
+# ---------------------------------------------------------------------------
+# resources/subscribe
+# ---------------------------------------------------------------------------
+
+
+def test_subscribe_result_valid_empty():
+    assert check_subscribe_result({}) == []
+
+
+def test_subscribe_result_not_dict():
+    checks = check_subscribe_result("nope")
+    ids = [c["id"] for c in checks]
+    assert "subscribe-result-type" in ids
+
+
+# ---------------------------------------------------------------------------
+# resources/unsubscribe
+# ---------------------------------------------------------------------------
+
+
+def test_unsubscribe_result_valid_empty():
+    assert check_unsubscribe_result({}) == []
+
+
+def test_unsubscribe_result_not_dict():
+    checks = check_unsubscribe_result(42)
+    ids = [c["id"] for c in checks]
+    assert "unsubscribe-result-type" in ids
+
+
+# ---------------------------------------------------------------------------
+# notifications/progress
+# ---------------------------------------------------------------------------
+
+
+def test_progress_notification_valid():
+    payload = {"params": {"progressToken": "tok-1", "progress": 50, "total": 100}}
+    assert check_progress_notification(payload) == []
+
+
+def test_progress_notification_integer_token():
+    payload = {"params": {"progressToken": 42, "progress": 0.5}}
+    assert check_progress_notification(payload) == []
+
+
+def test_progress_notification_params_not_dict():
+    checks = check_progress_notification({"params": "bad"})
+    ids = [c["id"] for c in checks]
+    assert "progress-params-type" in ids
+
+
+def test_progress_notification_missing_token():
+    checks = check_progress_notification({"params": {"progress": 1}})
+    ids = [c["id"] for c in checks]
+    assert "progress-token-missing" in ids
+
+
+def test_progress_notification_bool_token_rejected():
+    checks = check_progress_notification(
+        {"params": {"progressToken": True, "progress": 1}}
+    )
+    ids = [c["id"] for c in checks]
+    assert "progress-token-type" in ids
+
+
+def test_progress_notification_missing_progress():
+    checks = check_progress_notification({"params": {"progressToken": "t"}})
+    ids = [c["id"] for c in checks]
+    assert "progress-value-missing" in ids
+
+
+def test_progress_notification_bool_progress_rejected():
+    checks = check_progress_notification(
+        {"params": {"progressToken": "t", "progress": False}}
+    )
+    ids = [c["id"] for c in checks]
+    assert "progress-value-type" in ids
+
+
+def test_progress_notification_non_numeric_total():
+    checks = check_progress_notification(
+        {"params": {"progressToken": "t", "progress": 1, "total": "big"}}
+    )
+    ids = [c["id"] for c in checks]
+    assert "progress-total-type" in ids
+
+
+# ---------------------------------------------------------------------------
+# notifications/cancelled
+# ---------------------------------------------------------------------------
+
+
+def test_cancelled_notification_valid():
+    payload = {"params": {"requestId": "req-1", "reason": "user cancelled"}}
+    assert check_cancelled_notification(payload) == []
+
+
+def test_cancelled_notification_integer_request_id():
+    payload = {"params": {"requestId": 99}}
+    assert check_cancelled_notification(payload) == []
+
+
+def test_cancelled_notification_params_not_dict():
+    checks = check_cancelled_notification({"params": None})
+    ids = [c["id"] for c in checks]
+    assert "cancelled-params-type" in ids
+
+
+def test_cancelled_notification_missing_request_id():
+    checks = check_cancelled_notification({"params": {}})
+    ids = [c["id"] for c in checks]
+    assert "cancelled-request-id-missing" in ids
+
+
+def test_cancelled_notification_bool_request_id_rejected():
+    checks = check_cancelled_notification({"params": {"requestId": True}})
+    ids = [c["id"] for c in checks]
+    assert "cancelled-request-id-type" in ids
+
+
+def test_cancelled_notification_non_string_reason():
+    checks = check_cancelled_notification({"params": {"requestId": "r", "reason": 42}})
+    ids = [c["id"] for c in checks]
+    assert "cancelled-reason-type" in ids
+
+
+# ---------------------------------------------------------------------------
+# list_changed notifications (generic)
+# ---------------------------------------------------------------------------
+
+
+def test_list_changed_no_params():
+    assert check_list_changed_notification({}) == []
+
+
+def test_list_changed_empty_params():
+    assert check_list_changed_notification({"params": {}}) == []
+
+
+def test_list_changed_params_not_dict():
+    checks = check_list_changed_notification({"params": "bad"})
+    ids = [c["id"] for c in checks]
+    assert "list-changed-params-type" in ids
+
+
+# ---------------------------------------------------------------------------
+# notifications/resources/updated
+# ---------------------------------------------------------------------------
+
+
+def test_resources_updated_valid():
+    payload = {"params": {"uri": "file:///foo/bar"}}
+    assert check_resources_updated_notification(payload) == []
+
+
+def test_resources_updated_params_not_dict():
+    checks = check_resources_updated_notification({"params": 42})
+    ids = [c["id"] for c in checks]
+    assert "resources-updated-params-type" in ids
+
+
+def test_resources_updated_missing_uri():
+    checks = check_resources_updated_notification({"params": {}})
+    ids = [c["id"] for c in checks]
+    assert "resources-updated-uri-missing" in ids
+
+
+def test_resources_updated_empty_uri():
+    checks = check_resources_updated_notification({"params": {"uri": ""}})
+    ids = [c["id"] for c in checks]
+    assert "resources-updated-uri-missing" in ids
